Add a new subdirectory 'generic-ui' under middleware module.
